refactor(shareintent): extract URL boundary trimming into a pure unit
The trailing-trim logic (dropping punctuation and unbalanced closing
brackets from a greedily matched URL) is a URL-boundary concern distinct
from Intent parsing. Moving it to a pure, module-internal function names
that concept, gives it one responsibility, and lets it be unit-tested in
isolation instead of only end-to-end through the use case. Behavior is
unchanged; a reviewer flagged the O(N^2) trim as a potential ANR, but for
short share-intent text the impact is negligible, so the straightforward
implementation is kept and optimization is deferred.
