+++ app/src/main/java/io/github/omochice/pinosu/feature/shareintent/domain/usecase/ExtractSharedContentUseCaseImpl.kt
@@ -29,7 +29,7 @@ class ExtractSharedContentUseCaseImpl @Inject constructor() : ExtractSharedConte
 
     val urlMatch = URL_PATTERN.find(text)
     return if (urlMatch != null) {
-      val urlRange = urlMatch.range.first..trailingTrimmedEnd(text, urlMatch.range)
+      val urlRange = trimmedUrlRange(text, urlMatch.range)
       val url = text.substring(urlRange)
       val remaining = text.removeRange(urlRange).trim().replace(Regex("\\s{2,}"), " ")
       SharedContent(url = url, comment = remaining.ifBlank { null })
@@ -38,31 +38,7 @@ class ExtractSharedContentUseCaseImpl @Inject constructor() : ExtractSharedConte
     }
   }
 
-  /**
-   * Returns the inclusive end index of the URL within [range] after dropping trailing characters
-   * that are punctuation surrounding the URL rather than part of it. `\S+` greedily swallows a
-   * trailing `),.;` etc. (e.g. `"(https://example.com/a), ..."`), which would otherwise be saved as
-   * a broken URL. A closing bracket is only dropped when it is unbalanced within the match so that
-   * URLs legitimately containing balanced brackets are preserved.
-   */
-  private fun trailingTrimmedEnd(text: String, range: IntRange): Int {
-    var end = range.last
-    while (end > range.first && shouldDropTrailing(text[end], text, range.first, end)) {
-      end--
-    }
-    return end
-  }
-
-  private fun shouldDropTrailing(c: Char, text: String, start: Int, end: Int): Boolean =
-      c in TRAILING_PUNCTUATION ||
-          CLOSING_TO_OPENING[c]?.let { open ->
-            val sub = text.substring(start..end)
-            sub.count { it == open } < sub.count { it == c }
-          } ?: false
-
   companion object {
     private val URL_PATTERN = """https?://\S+""".toRegex(RegexOption.IGNORE_CASE)
-    private val TRAILING_PUNCTUATION = setOf('.', ',', ';', ':', '!', '?', '"', '\'')
-    private val CLOSING_TO_OPENING = mapOf(')' to '(', ']' to '[', '}' to '{')
   }
 }
